// src/integration/myairops/gen/booking.ts
export interface components {
    schemas: {
// ...
        "Airops.Flight.Booking.Models.QuotationViewModel": {
            /** Format: int32 */
            id?: number;
            quotationReference?: string | null;
            aircraftRegistration?: string | null;
            aircraftType?: string | null;
            /** @enum {string} */
            status?: "New" | "SentToCustomer" | "Booked" | "Invalidated";
            /** Format: double */
            totalGrossPrice?: number | null;
            currencyCode?: string | null;
            /** Format: date-time */
            createdDate?: string | null;
        };
// ...
        "Airops.Flight.Booking.Models.TripLegViewModel": {
            /** Format: int32 */
            readonly id?: number;
            /** @description third party identifier */
            externalReference?: string | null;
            /** Format: int32 */
            readonly tripId?: number;
            departureAirport: components["schemas"]["Airops.Flight.Booking.Models.TripLegAirportViewModel"];
            arrivalAirport: components["schemas"]["Airops.Flight.Booking.Models.TripLegAirportViewModel"];
            /**
             * Format: date-time
             * @description departure date/time in UTC. If not pinned, this will be calculated.
             *     If no legs have any date/times then the first sector departure time will default to the current date/time in UTC
             */
            departureDateTime?: string | null;
            /**
             * @description determines if the departure date/time is pinned
             * @default false
             */
            departureDateTimePinned: boolean;
            /**
             * Format: date-time
             * @description arrival date/time in UTC. If not pinned, this will be calculated.
             *     If no legs have any date/times then the first sector departure time will default to the current date/time in UTC
             */
            arrivalDateTime?: string | null;
            /**
             * @description determines if the arrival date/time is pinned
             * @default false
             */
            arrivalDateTimePinned: boolean;
            /**
             * Format: int32
             * @description flying time in minutes. If not pinned, this will be calculated
             */
            flyingTime: number;
            /**
             * @description determines if the flying time is pinned
             * @default false
             */
            flyingTimePinned: boolean;
            /**
             * Format: int32
             * @description block time in minutes. Calculated field.
             */
            readonly blockTime?: number;
            /**
             * Format: int32
             * @description number of adults flying
             */
            adults?: number;
            /**
             * Format: int32
             * @description number of children flying
             */
            children?: number;
            /**
             * Format: int32
             * @description number of crew
             */
            crew?: number;
            /**
             * Format: int32
             * @description number of cabin crew including flight attendants
             */
            cabinCrew?: number;
            techStop?: boolean;
            /** @description determines if the fuel uplift is needed */
            isFuelUpliftRequired?: boolean;
            /**
             * Format: date-time
             * @description On Task DateTime UTC
             */
            taskOnUtc?: string | null;
            /**
             * Format: date-time
             * @description Off Task DateTime Utc
             */
            taskOffUtc?: string | null;
            /**
             * Format: int32
             * @description Mission Task Time in minutes calculated field
             */
            readonly missionTime?: number | null;
            deadheadTaxNote?: string | null;
            /** Format: int32 */
            deadheadTaxBusinessPaxCount?: number | null;
            /** Format: int32 */
            deadheadTaxPersonalPaxCount?: number | null;
            legType: components["schemas"]["Airops.Flight.Booking.Models.TripLegTypeViewModel"];
            bookedAircraft?: components["schemas"]["Airops.Flight.Booking.Models.TripLegBookedAircraftViewModel"];
            requestorContactIds?: number[] | null;
            approvers?: components["schemas"]["Airops.Flight.Booking.Models.TripLegApproverViewModel"][] | null;
            billerContactIds?: number[] | null;
        };
// ...
        "Airops.Flight.Booking.Models.TripPassengerViewModel": {
            /** Format: int32 */
            id: number;
            /** Format: int32 */
            readonly tripId?: number;
            /** Format: int32 */
            crmPassengerId?: number | null;
            employeeId?: string | null;
            tripReason?: string | null;
            costCentre?: string | null;
            sharedSeat?: boolean;
            /** @description Only one passenger can be lead per trip. If set to true, the flag will be set to false on all other passengers. */
            leadPassenger?: boolean;
            hasAllergy?: boolean;
            details: components["schemas"]["Airops.Flight.Booking.Models.TripPassengerDetailViewModel"];
            address?: components["schemas"]["Airops.Flight.Booking.Models.TripPassengerAddressViewModel"];
            idDocument?: components["schemas"]["Airops.Flight.Booking.Models.TripPassengerIdDocumentViewModel"];
            visa?: components["schemas"]["Airops.Flight.Booking.Models.TripPassengerVisaViewModel"];
            apisAddress?: components["schemas"]["Airops.Flight.Booking.Models.TripPassengerAddressViewModel"];
        };
// ...
        "Airops.Flight.Booking.Models.TripWithLegsViewModel": {
            /** Format: int32 */
            readonly id?: number;
            /** @description myairops flight generated reference */
            readonly reference?: string | null;
            /** @description third party identifier */
            externalReference?: string | null;
            /** Format: int32 */
            clientId: number;
            /** Format: int32 */
            contactId?: number | null;
            /**
             * @description ICAO code
             * @example BE30
             */
            aircraftType?: string | null;
            /**
             * @description registration
             * @example GMY01
             */
            aircraft?: string | null;
            /**
             * @example TurboProp
             * @enum {string|null}
             */
            aircraftCategory?: "HelicopterSingleEngine" | "HelicopterTwinEngine" | "Piston" | "TurboProp" | "EntryLevelJet" | "LightJet" | "SuperLightJet" | "MidsizeJet" | "SuperMidsizeJet" | "HeavyJet" | "UltraLongRange" | "VIPAirliner" | "Airliner" | "eVTOL" | null;
            /** @enum {string|null} */
            readonly status?: "New" | "InProgress" | "Booked" | "Cancelled" | "Completed" | "Invoiced" | "Paid" | "Hold" | "CheckFeasibility" | null;
            notes?: components["schemas"]["Airops.Flight.Booking.Models.TripNotesViewModel"];
            tripType?: components["schemas"]["Airops.Flight.Booking.Models.TripLegTypeViewModel"];
            /**
             * @description defaults to Integration if not supplied. Can only be set on trip creation.
             * @enum {string}
             */
            source?: "Phone" | "Email" | "SocialMedia" | "Website" | "Marketplace" | "Integration" | "Other" | "Avinode" | "FlyEasy" | "Tuvoli" | "Aeroplan" | "Portal" | "BrokerReferral" | "ShuttleCustomer" | "EmployeeReferral" | "ClientPortal" | "PrivateFly" | "SalesPerson" | "NbaaAirmail";
            /** @description true if a quotation is required in order to book the trip */
            requiresQuotation?: boolean;
            /**
             * Format: int32
             * @description required if RequiresQuotation is set to true
             */
            bookedQuotationId?: number | null;
            approver?: string | null;
            /** Format: int32 */
            salesAgentId?: number | null;
            /** Format: int32 */
            salesAgentContactId?: number | null;
            externalMarketplaceReference?: string | null;
            legs: components["schemas"]["Airops.Flight.Booking.Models.TripLegViewModel"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}